*修改了GrapX在64位下的编译警告 *更换cpu构架宏和操作系统宏 *编译通过了vs2015 clstd安卓工程 +简单使用C++ 11实现了多线程模块，没有测试

+++ GrapX/Controls/GXUICtrlBase.cpp
@@ -577,7 +577,7 @@ namespace DlgXM
 
   GXINT GetPixelSizeFromMarkW(const GXWCHAR* szString)
   {
-    int nLen = GXSTRLEN(szString);
+    int nLen = (int)GXSTRLEN(szString);
     int nValue = GXATOI(szString);
     if(szString[nLen - 2] == 'p' && szString[nLen - 1] == 'x') {
       return nValue;
